Updating the jest configuration and TxSummary test
