+++ .eslintrc.js
@@ -22,7 +22,7 @@ module.exports = {
     {
       files: [
         '**/__tests__/*.{j,t}s?(x)',
-        '**/tests/unit/**/*.spec.{j,t}s?(x)',
+        '**/tests2/unit/**/*.spec.{j,t}s?(x)',
       ],
       env: {
         mocha: true,
