+++ .gitignore
@@ -2,7 +2,7 @@
 node_modules
 /dist
 
-/tests/e2e/reports/
+/tests2/e2e/reports/
 selenium-debug.log
 chromedriver.log
 geckodriver.log
