+++ package.json
@@ -5,13 +5,11 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
+    "test:unit": "vue-cli-service test:unit --setupTestFrameworkScriptFile=./tests/setup.js",
     "lint": "vue-cli-service lint",
     "scanner": "npx sonar-scanner"
   },
   "dependencies": {
-    "@babel/core": "^7.17.10",
-    "@babel/preset-env": "^7.17.10",
     "@ledgerhq/hw-app-btc": "^5.51.1",
     "@ledgerhq/hw-transport-webusb": "^5.46.0",
     "@rsksmart/rlogin": "^1.3.0",
@@ -41,27 +39,24 @@
     "web3-eth-contract": "^1.7.0"
   },
   "devDependencies": {
-    "@babel/preset-typescript": "^7.16.7",
     "@types/chai": "^4.2.11",
+    "@types/jest": "^27.5.0",
     "@types/node": "^15.3.0",
     "@types/sinon": "^10.0.2",
     "@types/w3c-web-usb": "^1.0.4",
     "@typescript-eslint/eslint-plugin": "^4.18.0",
     "@typescript-eslint/parser": "^4.18.0",
-    "@vue/babel-preset-app": "^5.0.4",
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-e2e-nightwatch": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-plugin-router": "~4.5.0",
     "@vue/cli-plugin-typescript": "~4.5.0",
-    "@vue/cli-plugin-unit-jest": "^5.0.4",
+    "@vue/cli-plugin-unit-jest": "^4.5.15",
     "@vue/cli-plugin-vuex": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
     "@vue/eslint-config-airbnb": "^5.0.2",
     "@vue/eslint-config-typescript": "^7.0.0",
-    "@vue/test-utils": "^1.3.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^8.2.5",
+    "@vue/test-utils": "^1.0.3",
     "chai": "^4.3.4",
     "chromedriver": "^93.0.1",
     "electron": "^15.0.0",
@@ -74,24 +69,10 @@
     "sass-loader": "^10.2.0",
     "sinon": "^11.1.2",
     "sonarqube-scanner": "^2.8.0",
-    "ts-jest": "^27.1.4",
     "typescript": "~4.1.5",
     "vue-cli-plugin-vuetify": "^2.4.5",
-    "vue-jest": "^3.0.7",
     "vue-template-compiler": "^2.6.11",
     "vuetify-loader": "^1.3.0",
     "webpack": "^4.46.0"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "vue"
-    ],
-    "transform": {
-      "^.+\\.ts$": "ts-jest",
-      "^.+\\.js$": "babel-jest",
-      ".*\\.vue$": "<rootDir>/node_modules/vue-jest"
-    }
   }
 }
