+++ tests/setup.js
@@ -1,0 +1,5 @@
+import Vue from 'vue';
+import Vuetify from 'vuetify';
+
+Vue.use(Vuetify);
+Vue.config.productionTip = false;
